requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity==1.13.0b4
Flask==2.3.2
openai==0.27.7
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0